Move DoubleEndedIterator to own module
